requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
opencv-python-headless>=4.9.0
ultralytics>=8.2.0
aiosqlite>=0.20.0
PyYAML>=6.0.1
numpy>=1.24.0
